feat: enable overriding embedded templates via the filesystem (#1)
Previously the only way to change the contents of the landing page was
to edit the template, recompile the `device-portal` binary, and tag a
new version of device-portal in order for deployment via Forklift. Now,
we can just use the same binary but change some template files in a
Forklift package which override the template embedded in
`device-portal`, removing the need to recompile and re-release
device-portal when changing the landing page.

This work is tracked in
https://www.notion.so/Allow-the-landing-page-to-load-template-files-from-the-filesystem-2754e612c78a80eea4bcfcff73ae815a?source=copy_link